Actually lock canvas in NFAtoDFA task
For some reason, importing an automaton respects a locked canvas, except for
the labels, which are still clickable. This commit fixes this problem in
the most stupid way possible, namely by relocking the canvas after the
automaton has been imported. We should probably fix the underlying
problem at some point.