Completely lock down canvas in lockCanvas()
There have been multiple attempts at disallowing the user any interaction with
the canvas when it is locked down by lockCanvas(). In order to complete this
work, we now simply lock down the whole canvas. Since all elements are children
of this element, this should not allow the user any interaction and serve only
as a display.