Redirect to ProblemSet index after deleting a problem
For some reason, reloading the overview of a problem set directly after
deleting a problem from it does not reflect the change. If the user goes back
to the list of problem sets and enters the edit-view from there, however, it
does show the change. As a hotfix, this commit redirects the user to the index
of problem sets after deleting a problem from a set instead.